refactor(modularity): Phase 1c - eliminate execution + adapters + services cross-module re-exports
Remove 4 cross-module re-export symbols:
- ExecutionRequest, ExecutionLaunchResult from execution/__init__.py (zero consumers)
- AdapterManifest from adapters/__init__.py (zero consumers)
- ConventionResolver from services/__init__.py (3 consumers)

Redirect ConventionResolver imports to source module (vibe3.config):
- execution/role_request_factory.py
- commands/task.py
- domain/qualify_gate.py

Note on adapters/__init__.py:
- Kept deep import 'from vibe3.models.adapter_manifest import AdapterManifest'
- Removed from __all__ to achieve public API goal
- Reason: mypy cannot resolve AdapterManifest through models lazy loader
  (not in TYPE_CHECKING block)

This completes Phase 1c of the modularity refactoring (#2276).

Tests: All modularity tests pass, full regression suite passes (3016 passed).
