+++ src/vibe3/execution/role_request_factory.py
@@ -5,12 +5,12 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
+from vibe3.config import get_convention
 from vibe3.execution.issue_role_support import (
     build_issue_async_cli_request,
     build_issue_sync_prompt_request,
 )
 from vibe3.models import ExecutionRequest, IssueInfo, OrchestraConfig
-from vibe3.services import get_convention
 
 
 def build_role_async_request(
